Add educational-program service

+++ api.schedule/schedule-educational-program-service/cmd/app/internal/model/speciality.go
@@ -1,8 +1,8 @@
 package model
 
 type Speciality struct {
-	ID        int64
+	ID        int
 	Name      string
 	Code      string
-	TermCount int64
+	TermCount int
 }
